Refactorizar MicrohponeButtonn, RecomendacionesInvalidass, RecommendationnsCarousel, SpeechBotCard, MicrophoneButton, TextField
