+++ DREAM Lab Frontend/src/pages/HomePage/components/RecommendationsCarousel/RecommendationsCarousel.jsx
