+++ DREAM Lab Frontend/src/pages/HomePage/components/SpeechBotCard/SpeechBotCard.css
@@ -31,12 +31,6 @@
     font-size: 20px;
 }
 
-.microphone-button {
-    font-size: 2.5em;
-    padding-right: 60%;
-    padding-left: 30%;
-}
-
 .send-button {
     width: 2rem;
     margin-left: 0.5rem; /* Adjust margin */
@@ -61,21 +55,4 @@
     .send-button {
         width: 2rem;
     }
-}
-
-/* Animación micrófono activo */
-@keyframes pulse {
-    0% {
-        transform: scale(1);
-    }
-    50% {
-        transform: scale(0.8);
-    }
-    100% {
-        transform: scale(1);
-    }
-}
-
-.relative.active svg {
-    animation: pulse 0.5s infinite alternate;
 }
