+++ DREAM Lab Frontend/src/pages/HomePage/components/SpeechBotCard/components/MicrophoneButton/MicrophoneButton.css
@@ -1,0 +1,29 @@
+.microphone-button {
+    padding-right: 60%;
+    padding-left: 30%;
+    font-size: 2.5em;
+}
+
+@media only screen and (max-width: 700px) {
+    .microphone-button {
+        justify-content: center;
+        font-size: 2.2rem;
+    }
+}
+
+/* Animación micrófono activo */
+@keyframes pulse {
+    0% {
+        transform: scale(1);
+    }
+    50% {
+        transform: scale(0.8);
+    }
+    100% {
+        transform: scale(1);
+    }
+}
+
+.relative.active svg {
+    animation: pulse 0.5s infinite alternate;
+}
